+++ src/primitive/mod.rs
@@ -630,6 +630,7 @@ impl ImplPrimitive {
             ImplPrimitive::Last => env.monadic_env(Value::last)?,
             ImplPrimitive::InverseBits => env.monadic_ref_env(Value::inverse_bits)?,
             ImplPrimitive::Unpartition => loops::unpartition(env)?,
+            ImplPrimitive::Ungroup => loops::ungroup(env)?,
             ImplPrimitive::InvTrace => trace(env, true)?,
         }
         Ok(())